requirements.txt
absl-py==2.1.0
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
backcall==0.2.0
batch-face==1.5.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.11
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.12.0
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
markupsafe==2.1.5
mistune==3.0.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.3.1
notebook-shim==0.2.4
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-transforms==0.0.6
opt-einsum==3.4.0
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.0.0
proglog==0.1.10
prometheus-client==0.21.1
prompt-toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydub==0.25.1
pyfacer==0.0.5
pygments==2.18.0
pyparsing==3.1.4
pysndfx==0.3.6
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pywavelets==1.4.1
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
send2trash==1.8.3
simplejson==3.19.3
six==1.17.0
sixdrepnet==0.1.6
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.7
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.5
tinycss2==1.4.0
tomli==2.2.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.3.1
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
validators==0.34.0
wave==0.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
widgetsnbextension==4.0.13
yacs==0.1.8
zipp==3.20.2